+++ src/main/java/com/github/cc3002/finalreality/controller/handlers/EnemyOutHandler.java
@@ -1,10 +1,11 @@
-package com.github.cc3002.finalreality.controller;
+package com.github.cc3002.finalreality.controller.handlers;
 
+import com.github.cc3002.finalreality.controller.GameController;
 import com.github.cc3002.finalreality.model.character.Enemy;
 
 import java.beans.PropertyChangeEvent;
 
-public class EnemyOutHandler implements IEventHandler{
+public class EnemyOutHandler implements IEventHandler {
     private final GameController gameController;
 
     public EnemyOutHandler(GameController gameController) {