+++ src/main/java/com/github/cc3002/finalreality/controller/handlers/PlayerCharacterOutHandler.java
@@ -1,10 +1,11 @@
-package com.github.cc3002.finalreality.controller;
+package com.github.cc3002.finalreality.controller.handlers;
 
+import com.github.cc3002.finalreality.controller.GameController;
 import com.github.cc3002.finalreality.model.character.player.IPlayerCharacter;
 
 import java.beans.PropertyChangeEvent;
 
-public class PlayerCharacterOutHandler implements IEventHandler{
+public class PlayerCharacterOutHandler implements IEventHandler {
     private final GameController gameController;
 
     public PlayerCharacterOutHandler(GameController gameController) {