+++ src/main/java/com/github/cc3002/finalreality/controller/phases/InvalidMovementException.java
@@ -1,0 +1,9 @@
+package com.github.cc3002.finalreality.controller.phases;
+/**
+ * @author Camila Labarca.
+ */
+public class InvalidMovementException extends Exception {
+    public InvalidMovementException(final String message) {
+        super(message);
+    }
+}