+++ src/main/java/com/github/cc3002/finalreality/controller/phases/InvalidTransitionException.java
@@ -1,0 +1,9 @@
+package com.github.cc3002.finalreality.controller.phases;
+/**
+ * @author Camila Labarca.
+ */
+public class InvalidTransitionException extends Exception{
+    public InvalidTransitionException(String message) {
+        super(message);
+    }
+}