+++ src/main/java/com/github/cc3002/finalreality/model/character/player/BlackMage.java
@@ -1,11 +1,8 @@
 package com.github.cc3002.finalreality.model.character.player;
 
-import com.github.cc3002.finalreality.model.character.ICharacter;
 import com.github.cc3002.finalreality.model.weapon.IWeapon;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
-import java.util.concurrent.BlockingQueue;
 
 public class BlackMage extends PlayerCharacter{
 
@@ -16,11 +13,10 @@ public class BlackMage extends PlayerCharacter{
      * @param name       the mage character's name
      * @param points     the mage character's points
      * @param defense    the mage character's defense
-     * @param turnsQueue the queue with the characters waiting for their turn
      * @param mana       the mage character's mana
      */
-    public BlackMage(String name, int points, int defense, @NotNull BlockingQueue<ICharacter> turnsQueue, int mana) {
-        super(name, points, defense, turnsQueue);
+    public BlackMage(String name, int points, int defense, int mana) {
+        super(name, points, defense);
         this.mana = mana;
     }