+++ src/test/java/com/github/cc3002/finalreality/model/character/AbstractCharacterTest.java
@@ -1,9 +1,5 @@
 package com.github.cc3002.finalreality.model.character;
 
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
-import org.junit.jupiter.api.Assertions;
-
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
@@ -16,28 +12,9 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public abstract class AbstractCharacterTest {
 
-  protected BlockingQueue<ICharacter> turns;
-
   /**
-   * Checks that the character waits the appropriate amount of time for it's turn.
+   * Checks the construction of a character
    */
-  protected void checkWaitTurn(ICharacter character) {
-    Assertions.assertTrue(turns.isEmpty());
-    character.waitTurn();
-    try {
-      // Thread.sleep is not accurate so this values may be changed to adjust the
-      // acceptable error margin.
-      // We're testing that the character waits approximately 1 second.
-      Thread.sleep(900);
-      Assertions.assertEquals(0, turns.size());
-      Thread.sleep(200);
-      Assertions.assertEquals(1, turns.size());
-      Assertions.assertEquals(character, turns.peek());
-    } catch (InterruptedException e) {
-      e.printStackTrace();
-    }
-  }
-
   protected void checkConstruction(final ICharacter expectedCharacter,
                                    final ICharacter testEqualCharacter,
                                    final ICharacter sameClassDifferentCharacter,
@@ -46,13 +23,12 @@ public abstract class AbstractCharacterTest {
     assertNotEquals(sameClassDifferentCharacter, testEqualCharacter);
     assertNotEquals(testEqualCharacter, differentClassCharacter);
     assertEquals(expectedCharacter.hashCode(), testEqualCharacter.hashCode());
+    assertEquals(expectedCharacter.getDefense(), testEqualCharacter.getDefense());
   }
 
-
-  protected void basicSetUp() {
-    turns = new LinkedBlockingQueue<>();
-  }
-
+  /**
+   * Checks a character's attack
+   */
   protected void checkAttack(ICharacter attacker, ICharacter attacked){
     assertTrue(attacker.isAlive());
     assertTrue(attacked.isAlive());