+++ src/test/java/com/github/cc3002/finalreality/model/character/EngineerTest.java
@@ -22,51 +22,66 @@ public class EngineerTest extends AbstractPlayerCharacterTest {
 
     @BeforeEach
     void setUp() {
-        super.basicSetUp();
-        vivi = new Engineer(ENGINEER_NAME, 10,10,turns);
+        vivi = new Engineer(ENGINEER_NAME, 10,10);
     }
 
-    @Test
-    void waitTurnTest() {
-        vivi.equip(axe);
-        checkWaitTurn(vivi);
-    }
-
+    /**
+     * Checks the construction of a engineer
+     */
     @Test
     void constructorTest() {
-        checkConstruction(new Engineer(ENGINEER_NAME, 10, 10, turns),
+        checkConstruction(new Engineer(ENGINEER_NAME, 10, 10),
                 vivi,
-                new Engineer("Different name",10, 10, turns),
-                new Knight(ENGINEER_NAME,10, 10, turns));
+                new Engineer("Different name",10, 10),
+                new Knight(ENGINEER_NAME,10, 10));
     }
 
-
+    /**
+     * Checks it doesn't equip a sword
+     */
     @Test
     void equippedSwordTest() {
         checkNotEquippedWeapon(vivi, sword);
     }
+
+    /**
+     * Checks it equips an axe
+     */
     @Test
     void equippedAxeTest() {
         checkEquippedWeapon(vivi, axe);
     }
+
+    /**
+     * Checks it doesn't equip a knife
+     */
     @Test
     void equippedKnifeTest() {
         checkNotEquippedWeapon(vivi, knife);
     }
 
+    /**
+     * Checks it doesn't equip a staff
+     */
     @Test
     void equippedStaffTest() {
         checkNotEquippedWeapon(vivi, staff);
     }
 
+    /**
+     * Checks it equips a bow
+     */
     @Test
     void equippedBowTest() {
         checkEquippedWeapon(vivi, bow);
     }
 
+    /**
+     * Checks a dead engineer doesn't equip
+     */
     @Test
     void equipDeadEngineer(){
-        Enemy enemy = new Enemy("Enemy", 10, 10, 10, 20, turns);
+        Enemy enemy = new Enemy("Enemy", 10, 10, 10, 20);
         enemy.attack(vivi);
         vivi.equip(bow);
         assertNull(vivi.getEquippedWeapon());